backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sqlalchemy
alembic
httpx
requests
google-genai
python-telegram-bot
cloudinary
